// src/redux/contract/types.ts
export enum ActionType {
  LOAD_CURRENT_CONTRACT = 'CONTRACT/LOAD_CURRENT_CONTRACT',
  LOAD_CURRENT_CONTRACT_SUCCESS = 'CONTRACT/LOAD_CURRENT_CONTRACT_SUCCESS',
  LOAD_CONTRACT_SEARCH_LIST = 'CONTRACT/LOAD_CONTRACT_SEARCH_LIST',
  LOAD_CONTRACT_SEARCH_LIST_SUCCESS = 'CONTRACT/LOAD_CONTRACT_SEARCH_LIST_SUCCESS',
  LOAD_CONTRACT_LIST = 'CONTRACT/LOAD_CONTRACT_LIST',
  LOAD_CONTRACT_LIST_SUCCESS = 'CONTRACT/LOAD_CONTRACT_LIST_SUCCESS',
  LOAD_CONTRACT_LOADING = 'CONTRACT/LOAD_CONTRACT_LOADING',
  CHANGE_LOADING_STATE = 'CONTRACT/CHANGE_LOADING_STATE',
  END_EDIT = 'CONTRACT/END_EDIT',
  LOGOUT = 'CONTRACT/LOGOUT',
}